+++ README.md
@@ -11,10 +11,8 @@ Executable in local if provided a valid bot token
 TODOs:
 [TECH] :
 Naming improvements\
-Evaluate creating a proper formal structure pattern to represent user commands & args\
 Get to 100% uptime (heroku wakeup)\
 think whether using action for non-secret env related conf is best
-don't panic on command error
 
 [FUNCTIONAL]
 Impl more commands: score mgt, rio check\
