+++ src/main/kotlin/com/lmo/ninie/io/models/SongService.kt
@@ -1,4 +1,4 @@
-package com.lmo.ninie.io.services.models
+package com.lmo.ninie.io.models
 
 import com.lmo.ninie.io.models.songs.Song
 import io.vavr.control.Option
